bench(json): Include `winnow` in the cross-parser compare
`winnow` seems to be taking half the time of `nom` and I had to check
the output to confirm there wasn't a bug.  I dumped the `JsonZero`
structure from both parsers and diffed it and it was the same.

This was a direct port, using the existing `nom` pieces.  There are
parts more specialized to `winnow` that should speed it up.