docs: link examples & tutorial of right version on docs.rs
Previously the start page of the documentation on docs.rs for chumsky
linked the examples and tutorial from the main branch, no matter
which version you selected on docs.rs.

So e.g. https://docs.rs/chumsky/0.9.2/chumsky/ links
https://github.com/zesterer/chumsky/blob/main/examples/brainfuck.rs
which is confusing since the main branch currently contains the source
code for the upcoming 1.0.0, which has a different API.

This commit remedies this for future releases by separating the link labels
and URLs using CommonMark link reference definitions and overriding[1]
them in src/lib.rs, building the correct URLs via CARGO_PKG_VERSION[2].

This trick will require git tags to be created for all future versions
(so e.g. if 0.9.3 will be released the commit will have to be tagged
with "0.9.3").

[1]: https://spec.commonmark.org/0.30/#example-204
[2]: https://doc.rust-lang.org/cargo/reference/environment-variables.html