Implement the None associativity
The infix functions need to return an InfixResult to represent
the case when parsing is ambigious due to no associtivity.

The parse function will now also return an error if an infix parser
inside it's loop return an ambigious result.

In order to correctly determinate if an infix parse is ambigious,
the operator must be parsed before the associativity is checked.
Otherwise a "probe" call by the vector implementaion might return
the Ambigious error even if this operator is not present.