Implement the None associativity
All pratt operators can now return two types of errors,
non-fatal (`NoMatch`), which allow the parse to continue and
fatal (`Err`) errors, that abort the parse.

If a non-associative infix operator encounters an operator with
the same binding strength it returns a fatal error because of
an ambigious input.

For this to work, the infix operator mus first parse it's operator
before checking the binding power.