Add non-associative infix operators (#800)
* Rename binding_power to precedence

Binding power is an implementation detail, what users specify
is the precedence level.

* Change binding power to i32

The binding power zero is used as a special marker for the first
invocation.
User can specify a precedence of zero resulting in a binding power
of zero, which would mimic this marker.

Use an i32 and i32:MIN as a marker instead.

* Add Associativity::None

This designates an infox operator as non-associative.
Chained operator result in a parse error because without an
associativity they can't be unambigiously parsed.

Also add some test cases that will fail for now.

* Implement the None associativity

All pratt operators can now return two types of errors,
non-fatal (`NoMatch`), which allow the parse to continue and
fatal (`Err`) errors, that abort the parse.

If a non-associative infix operator encounters an operator with
the same binding strength it returns a fatal error because of
an ambigious input.

For this to work, the infix operator mus first parse it's operator
before checking the binding power.

* Only check ambiguity with non-associative infix operators