fix(494): add another recursive parser
This commit (hopefully) renders `define` and `declare` for users of the
library obselete, by adding a parser which wraps the declaration and
definition similar to `recursive`. This `recursive_n` allows for the
definition of mutually recursive parsers, resulting in a single parser.
This solves the memory leak in a fashion similar to #664 by keeping all
strong references within `RecursiveN` and dropping these all whenever the
parser returned by `recursive_n` is dropped.

Additionally, `Recursive::declare` and `Recursive::define` are removed
from the API.